+++ server/router/api/v2/test/routeBuilder/RouteBuilder.js
@@ -1,12 +1,15 @@
-import {serializeReq} from "./serializeReq.js";
-import validate from "./validate.js";
-import {addController} from "./addController.js";
-import {serializeRes} from "./serializeRes.js";
+import {serializeReq} from "./core/pipelineHandlers/serializeReq.js";
+import validate from "./core/pipelineHandlers/validate.js";
+import {addController} from "./core/pipelineHandlers/addController.js";
+import {serializeRes} from "./core/pipelineHandlers/serializeRes.js";
 import {Router} from "express";
-import {APIError} from "../../../../../../util/error/APIError.js";
-import {ErrorReason} from "../../../../../../util/error/ErrorReason.js";
-import {Method} from "./Method.js";
-import {authenticate} from "./authenticate.js";
+import {APIError} from "./error/APIError.js";
+import {ErrorReason} from "./error/ErrorReason.js";
+import {Method} from "./core/enums/Method.js";
+import {authenticate} from "./core/pipelineHandlers/authenticate.js";
+import {authorize} from "./core/pipelineHandlers/authorize.js";
+import isAllowed from "./core/authorization/isAllowed.js";
+import {Action} from "./core/enums/Action.js";
 
 export class RouteBuilder {
     /**
@@ -19,6 +22,7 @@ export class RouteBuilder {
         this.method = method;
 
         this.authenticator = null;
+        this.authorizer = null;
 
         this.reqSerializer = null;
         this.reqValidator = null;
@@ -29,6 +33,13 @@ export class RouteBuilder {
 
     authenticate = function (){
         this.authenticator = authenticate;
+        return this;
+    }
+
+    authorize = function (resource, action=null){
+        this.authenticator = authenticate;
+        const actionToAuthorize = action ?? determineAction(this.method);
+        this.authorizer = authorize(actionToAuthorize, resource, isAllowed);
         return this;
     }
 
@@ -85,11 +96,24 @@ export class RouteBuilder {
     }
 
     #addPipeConfigToRouter = function (router){
-        let pipeHandlersToApply = [this.authenticator, this.reqSerializer, this.reqValidator, this.controller, this.resSerializer];
+        let pipeHandlersToApply = [this.authenticator, this.authorizer, this.reqSerializer, this.reqValidator, this.controller, this.resSerializer];
         for(let i=0, len=pipeHandlersToApply.length; i<len; i++)
             pipeHandlersToApply[i] = pipeHandlersToApply[i] ?? this.#pipeHandlerMocker;
 
         return router[this.method](this.endpoint, ...pipeHandlersToApply);
     }
     #pipeHandlerMocker = function (req, res, next) { return next(); }
+}
+
+function determineAction(httpMethod) {
+    switch(httpMethod){
+        case Method.POST:
+            return Action.CREATE;
+        case Method.GET:
+            return Action.READ;
+        case Method.PUT:
+            return Action.UPDATE;
+        case Method.DELETE:
+            return Action.DELETE;
+    }
 }
